X11: Support for fractional scaling
On X11, we can support fractional scaling. Let's convert `scale` to a
float number and round it automatically on multiplication. When scale
is used, we `round()` the result.

+++ src/output.h
@@ -45,7 +45,7 @@ struct output {
         bool (*is_idle)(void);
         bool (*have_fullscreen_window)(void);
 
-        int (*get_scale)(void);
+        double (*get_scale)(void);
 };
 
 /**
